fixed size messages and reduced allocations

+++ include/order_simulator.hpp
@@ -7,13 +7,13 @@
 
 class OrderSimulator{
 	public:
-		OrderSimulator(double cp, uint64_t starting_id = 1);			
+		OrderSimulator(int64_t cp, uint64_t starting_id = 1);
 		Order generate_order();
 		std::vector<Order> generate_n_orders(int n);
 
 	private:
 		uint64_t id_counter_;
-		double center_price_;
+		int64_t center_price_;
 		std::mt19937 rng_; 
 
 };
